refactor(client): rename files to kebab-case
